Use ontology URI derivative for namespace
All other SPAR ontologies use a form of the namespace http://purl.org/spar for their namespace except this one. It uses http://purl.org/spar/po for the ontology URI but the namespace is the older essepuntato one. This just needs to be updated, as per this single-line update.